move classes, refactor
