+++ src/main/java/frc/robot/elevator/CoralRoller.java
@@ -1,4 +1,4 @@
-package frc.robot.grippers;
+package frc.robot.elevator;
 
 import com.revrobotics.RelativeEncoder;
 import com.revrobotics.spark.SparkBase.ControlType;
